PLEASE SEE EXTENDED EXPLANATION......
I modified both the BusStationMain and CityDefaultInitialization classes. I combined all other classes into the project and it works; however, Drake said that he will work on the CityDefaultInitialization class to fix the repeated routes. As long as it returns the array of three routes the program will run great. The output of the result in the GUI is just a fix for now. We can either keep it or change it later. There are issues with the GUI but doesn't hurt the functionality. Such as opening multiple GUI's when you should not be able to. I will fix those as well. I do have some error handling issues I have to take care of this weekend as well. Currently my sorter is broken and I have to fix that.