Update CityDefaultInitialization and 3 others
Up to three routes, this time guaranteed always to be different, are generated. If two stations lie in a straight line, only one route will be generated, and the other two will be null. If two stations are diagonally across from each other, only two routes will be generated, and the other one will be null.